+++ src/zoid/message/component.js
@@ -11,11 +11,14 @@ import {
     getLibraryVersion,
     runStats,
     logger,
-    getStorageID,
     getSessionID,
     getGlobalState,
     getCurrentTime,
-    getStageTag
+    writeStorageID,
+    getOrCreateStorageID,
+    getStageTag,
+    ppDebug,
+    isScriptBeingDestroyed
 } from '../../utils';
 import validate from './validation';
 import containerTemplate from './containerTemplate';
@@ -118,6 +121,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                             offer: offerType,
                             refId: messageRequestId,
                             refIndex: index,
+                            src: 'message_click',
                             onClose: () => focus()
                         });
 
@@ -126,6 +130,7 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                             et: 'CLICK',
                             event_type: 'MORS'
                         });
+
                         logger.track({
                             index,
                             et: 'CLICK',
@@ -170,9 +175,13 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 value: ({ props }) => {
                     const { onReady } = props;
 
-                    return ({ meta, activeTags }) => {
+                    return ({ meta, activeTags, deviceID }) => {
                         const { account, merchantId, index, modal, getContainer } = props;
-                        const { messageRequestId, displayedMessage, trackingDetails, offerType } = meta;
+                        const { messageRequestId, trackingDetails, offerType, ppDebugId } = meta;
+                        ppDebug(`Message Correlation ID: ${ppDebugId}`);
+
+                        // Write deviceID from iframe localStorage to merchant domain localStorage
+                        writeStorageID(deviceID);
 
                         logger.addMetaBuilder(existingMeta => {
                             // Remove potential existing meta info
@@ -180,12 +189,22 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                             // eslint-disable-next-line no-param-reassign
                             delete existingMeta[index];
 
+                            // Need to capture existing attributes under global before destroying
+                            const { global: existingGlobal = {} } = existingMeta;
+                            // eslint-disable-next-line no-param-reassign
+                            delete existingMeta.global;
+
                             return {
+                                // Need to merge global attribute here due to preserve performance attributes
+                                global: {
+                                    ...existingGlobal,
+                                    deviceID, // deviceID from internal iframe storage
+                                    sessionID: getSessionID() // Session ID from parent local storage
+                                },
                                 [index]: {
                                     type: 'message',
                                     messageRequestId,
                                     account: merchantId || account,
-                                    displayedMessage,
                                     trackingDetails
                                 }
                             };
@@ -252,7 +271,11 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                         const container = getContainer();
                         // Let the cleanup finish before re-rendering
                         ZalgoPromise.delay(0).then(() => {
-                            if (container && container.ownerDocument.body.contains(container)) {
+                            if (
+                                container &&
+                                container.ownerDocument.body.contains(container) &&
+                                !isScriptBeingDestroyed()
+                            ) {
                                 // Will re-render with the full config options stored in the zoid props
                                 const { render, state, updateProps, clone } = messagesMap.get(container).clone();
 
@@ -292,32 +315,49 @@ export default createGlobalVariableGetter('__paypal_credit_message__', () =>
                 queryParam: true,
                 sendToChild: false,
                 required: false,
-                value: getMeta
+                value: getMeta,
+                debug: ppDebug(`SDK Meta: ${getMeta()}`)
             },
             env: {
                 type: 'string',
                 queryParam: true,
-                value: getEnv
+                value: getEnv,
+                debug: ppDebug(`Environment: ${getEnv()}`)
             },
             version: {
                 type: 'string',
                 queryParam: true,
-                value: getLibraryVersion
+                value: getLibraryVersion,
+                debug: ppDebug(`Library Version: ${getLibraryVersion()}`)
             },
             deviceID: {
                 type: 'string',
                 queryParam: true,
-                value: getStorageID
+                value: getOrCreateStorageID,
+                debug: ppDebug(`Device ID: ${getOrCreateStorageID()}`)
             },
             sessionID: {
                 type: 'string',
                 queryParam: true,
-                value: getSessionID
+                value: getSessionID,
+                debug: ppDebug(`Session ID: ${getSessionID()}`)
             },
             scriptUID: {
                 type: 'string',
                 queryParam: true,
-                value: getCurrentScriptUID
+                value: getCurrentScriptUID,
+                debug: ppDebug(`ScriptUID: ${getCurrentScriptUID()}`)
+            },
+            debug: {
+                type: 'boolean',
+                queryParam: 'pp_debug',
+                value: () => /(\?|&)pp_debug=true(&|$)/.test(window.location.search)
+            },
+            messageLocation: {
+                type: 'string',
+                queryParam: false,
+                value: () => window.location.href,
+                debug: ppDebug(`Message Location: ${window.location.href}`)
             },
             stageTag: {
                 type: 'string',