+++ utils/devServerProxy/proxy.js
@@ -6,6 +6,9 @@ import { PORT, VARIANT } from '../../server/constants';
 import { populateTemplate, localizeCurrency } from './miscellaneous';
 import { getTerms } from './mockTerms';
 
+// set this environment variable to simulate the time for the request to be answered
+const requestDelay = process.env.REQUEST_DELAY ?? 500;
+
 const devAccountMap = {
     DEV00000000NI: ['US', ['ni'], 'ni'],
     DEV0000000NIQ: ['US', ['ni'], 'niq'],
@@ -29,6 +32,14 @@ const devAccountMap = {
     DEV0000000IAG: ['DE', ['inst'], 'inst_any_gtz'],
     DEV000000PQAG: ['DE', ['inst'], 'palaq_any_gtz'],
     DEV000000PQAZ: ['DE', ['inst'], 'palaq_any_eqz'],
+    DEV000DEPLEQZ: ['DE', ['gpl'], 'gpl_eqz'],
+    DEV000DEPLGTZ: ['DE', ['gpl'], 'gpl_gtz'],
+    DEV00DEPLQEQZ: ['DE', ['gpl'], 'gplq_eqz'],
+    DEV00DEPLQGTZ: ['DE', ['gpl'], 'gplq_gtz'],
+    DEV0XBDEPLEQZ: ['DE', ['gpl'], 'gpl_eqz-non-de'],
+    DEV0XBDEPLGTZ: ['DE', ['gpl'], 'gpl_gtz-non-de'],
+    DEVXBDEPLQEQZ: ['DE', ['gpl'], 'gplq_eqz-non-de'],
+    DEVXBDEPLQGTZ: ['DE', ['gpl'], 'gplq_gtz-non-de'],
 
     DEV000000GBPL: ['GB', ['gpl'], 'pl'],
     DEV00000GBPLQ: ['GB', ['gpl'], 'plq'],
@@ -199,9 +210,60 @@ export default (app, server, compiler) => {
         const account = clientId || payerId || merchantId;
         const [country, productNames] = devAccountMap[account] ?? ['US', ['ni']];
 
-        const productsJSON = productNames.map(product =>
-            fs.readFileSync(`modals/${country}/${product}.json`, 'utf-8').toString()
-        );
+        const productsJSON = productNames.map(product => {
+            const jsonFile = `modals/${country}/${product}.json`;
+            if (fs.existsSync(jsonFile)) {
+                return fs.readFileSync(jsonFile, 'utf-8').toString();
+            }
+            // eslint-disable-next-line no-console
+            console.warn(`${jsonFile} does not exist`);
+            return `{
+                    "meta": {
+                        "product": "GPL",
+                        "periodicPayment": "{formattedPeriodicPayment}",
+                        "minAmount": "{minAmount}",
+                        "maxAmount": "{maxAmount}",
+                        "qualifying": "{qualifying_offer}",
+                        "amount": "{transaction_amount}",
+                        "variables": {
+                            "transaction_amount": "\${eval(transaction_amount ? transaction_amount : '-')}",
+                            "qualifying_offer": "\${eval(CREDIT_OFFERS_DS.qualifying_offer ? CREDIT_OFFERS_DS.qualifying_offer : 'false')}",
+                            "financing_code": "\${CREDIT_OFFERS_DS.financing_code}",
+                            "formattedPeriodicPayment": "\${CREDIT_OFFERS_DS.formattedPeriodicPayment}",
+                            "total_payments": "\${CREDIT_OFFERS_DS.total_payments}",
+                            "formattedMinAmount": "\${CREDIT_OFFERS_DS.formattedMinAmount}",
+                            "formattedMaxAmount": "\${CREDIT_OFFERS_DS.formattedMaxAmount}",
+                            "formattedTotalCost": "\${CREDIT_OFFERS_DS.formattedTotalCost}",
+                            "minAmount": "\${CREDIT_OFFERS_DS.minAmount}",
+                            "maxAmount": "\${CREDIT_OFFERS_DS.maxAmount}",
+                            "apr": "\${CREDIT_OFFERS_DS.apr}",
+                            "nominal_rate": "\${CREDIT_OFFERS_DS.nominal_rate}"
+                        }
+                    },
+                    "content": {
+                        "headline": {
+                            "singleProduct": "Pay in X"
+                        },
+                        "subHeadline": {
+                            "pay": {
+                                "start": "Make one interest-free payment",
+                                "amount": "of {formattedPeriodicPayment}",
+                                "end": "today, then pay the rest monthly."
+                            },
+                            "available": "Available on purchases from {formattedMinAmount} to {formattedMaxAmount}.",
+                            "apply": "Apply lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus, enim?."
+                        },
+                        "terms": [
+                            "TERMS AND CONDITIONS.",
+                            "PayPal Pay in X Lorem ipsum dolor sit, amet consectetur adipisicing, elit.",
+                            "Facilis adipisci quidem obcaecati, accusantium voluptatum repudiandae magni dicta officiis est eum!"
+                        ],
+                        "instructions": {
+                            "title": ["Check out securely with", "PayPal", "and choose", "Pay in X"]
+                        }
+                    }
+                }`;
+        });
 
         const terms = getTerms(country, Number(amount));
         const [bestOffer] = terms.offers || [{}];
@@ -268,23 +330,39 @@ export default (app, server, compiler) => {
         return { props, productNames };
     };
 
-    app.get('/credit-presentment/smart/modal', (req, res) => {
+    // create the initial modal with content
+    app.get('/credit-presentment/smart/modal', async (req, res) => {
         const { targetMeta, scriptUID } = req.query;
+
         const { props, productNames } = getModalData(req);
 
-        res.send(
-            createMockZoidMarkup(
-                targetMeta ? 'modal' : `modal-${productNames.includes('ezp_old') ? 'US-EZP' : props.country}`,
-                `<script>crc.setupModal(${JSON.stringify(props)})</script>`,
-                scriptUID
-            )
-        );
+        const component = () => {
+            if (props.country === 'US' && productNames.includes('ezp_old')) {
+                return 'US-EZP';
+            }
+
+            if (props.country === 'DE' && productNames.includes('gpl')) {
+                return 'DE-GPL';
+            }
+
+            return props.country;
+        };
+        setTimeout(() => {
+            res.send(
+                createMockZoidMarkup(
+                    targetMeta ? 'modal' : `modal-${component()}`,
+                    `<script>crc.setupModal(${JSON.stringify(props)})</script>`,
+                    scriptUID
+                )
+            );
+        }, requestDelay);
     });
-    app.get('/credit-presentment/modalContent', (req, res) => {
+
+    app.get('/credit-presentment/modalContent', async (req, res) => {
         const { props: data } = getModalData(req);
         setTimeout(() => {
             res.send(data);
-        }, 1000);
+        }, requestDelay);
     });
 
     app.get('/credit-presentment/renderMessage', async (req, res) => {
@@ -303,14 +381,14 @@ export default (app, server, compiler) => {
 
         setTimeout(() => {
             res.send(getTerms(country, Number(amount)));
-        }, 1000);
+        }, requestDelay);
     });
     app.get('/credit-presentment/calculateTerms', (req, res) => {
         const { country, amount } = req.query;
 
         setTimeout(() => {
             res.send(getTerms(country, Number(amount)));
-        }, 1000);
+        }, requestDelay);
     });
 
     app.get('/credit-presentment/messages', (req, res) => {